+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [13.9.2](https://kaos.sh/ek/13.9.2)
 
 - `[knf]` Added helper `Q`
+- `[fmtc]` Code refactoring
 
 ### [13.9.1](https://kaos.sh/ek/13.9.1)